Testing hardware rendering using legacy renderer
