+++ .vscode/settings.json
@@ -87,7 +87,8 @@
         "shared_mutex": "cpp",
         "cfenv": "cpp",
         "future": "cpp",
-        "valarray": "cpp"
+        "valarray": "cpp",
+        "csetjmp": "cpp"
     },
     "C_Cpp.default.cppStandard": "c++14",
     "C_Cpp.default.cStandard": "c11"
